refactor(packages): Grouped remaining classes into com.example.demo.model & com.example.demo.view packages
Updated module exports
