+++ src/javadocs/com.example.demo/com/example/demo/BossProjectile.html
@@ -87,10 +87,10 @@ loadScripts(document, 'script');</script>
 <div class="inheritance" title="Inheritance Tree"><a href="https://docs.oracle.com/en/java/javase/22/docs/api/java.base/java/lang/Object.html" title="class or interface in java.lang" class="external-link">java.lang.Object</a>
 <div class="inheritance">javafx.scene.Node
 <div class="inheritance">javafx.scene.image.ImageView
-<div class="inheritance"><a href="ActiveActor.html" title="class in com.example.demo">com.example.demo.ActiveActor</a>
-<div class="inheritance"><a href="ActiveActorDestructible.html" title="class in com.example.demo">com.example.demo.ActiveActorDestructible</a>
-<div class="inheritance"><a href="Projectile.html" title="class in com.example.demo">com.example.demo.Projectile</a>
-<div class="inheritance">com.example.demo.BossProjectile</div>
+<div class="inheritance"><a href="ActiveActor.html" title="class in com.example.demo">com.example.demo.model.ActiveActor</a>
+<div class="inheritance"><a href="ActiveActorDestructible.html" title="class in com.example.demo">com.example.demo.model.ActiveActorDestructible</a>
+<div class="inheritance"><a href="Projectile.html" title="class in com.example.demo">com.example.demo.model.Projectile</a>
+<div class="inheritance">com.example.demo.model.BossProjectile</div>
 </div>
 </div>
 </div>
