+++ src/javadocs/com.example.demo/com/example/demo/ShieldImage.html
@@ -87,7 +87,7 @@ loadScripts(document, 'script');</script>
 <div class="inheritance" title="Inheritance Tree"><a href="https://docs.oracle.com/en/java/javase/22/docs/api/java.base/java/lang/Object.html" title="class or interface in java.lang" class="external-link">java.lang.Object</a>
 <div class="inheritance">javafx.scene.Node
 <div class="inheritance">javafx.scene.image.ImageView
-<div class="inheritance">com.example.demo.ShieldImage</div>
+<div class="inheritance">com.example.demo.view.ShieldImage</div>
 </div>
 </div>
 </div>
