+++ src/main/java/com/example/demo/level/BossLevel.java
@@ -1,6 +1,6 @@
 package com.example.demo.level;
 
-import com.example.demo.Boss;
+import com.example.demo.model.Boss;
 import javafx.animation.KeyFrame;
 import javafx.animation.Timeline;
 import javafx.util.Duration;
