+++ src/main/java/com/example/demo/level/BossLevelView.java
@@ -1,6 +1,6 @@
 package com.example.demo.level;
 
-import com.example.demo.ShieldImage;
+import com.example.demo.view.ShieldImage;
 import javafx.scene.Group;
 
 public class BossLevelView extends LevelView {
