+++ src/main/java/com/example/demo/model/ActiveActor.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example.demo.model;
 
 import javafx.scene.image.*;
