+++ src/main/java/com/example/demo/model/Boss.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example.demo.model;
 
 import java.util.*;
