+++ src/main/java/com/example/demo/model/BossProjectile.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example.demo.model;
 
 public class BossProjectile extends Projectile {
