+++ src/main/java/com/example/demo/model/FighterPlane.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example.demo.model;
 
 public abstract class FighterPlane extends ActiveActorDestructible {
