+++ src/main/java/com/example/demo/model/Projectile.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example.demo.model;
 
 public abstract class Projectile extends ActiveActorDestructible {
