+++ src/main/java/com/example/demo/model/UserPlane.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example.demo.model;
 
 public class UserPlane extends FighterPlane {
