+++ src/main/java/com/example/demo/model/UserProjectile.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example.demo.model;
 
 public class UserProjectile extends Projectile {
