+++ src/main/java/com/example/demo/view/LevelSelection.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example.demo.view;
 
 import javafx.scene.Scene;
 import javafx.scene.image.Image;
