+++ src/main/java/com/example/demo/view/MainMenu.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example.demo.view;
 
 import com.example.demo.controller.Controller;
