+++ src/main/java/com/example/demo/view/ShieldImage.java
@@ -1,4 +1,4 @@
-package com.example.demo;
+package com.example.demo.view;
 
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
